share db code for scripts
want to pull this from ../src/models/gbdb eventually
